# Optimizations in lazy.py