… test/test_custom_function.py
… tinygrad/lazy.py