Before:

```
lazy.py: 0.9059 seconds
tensor.py: 0.4541 seconds
shapetracker.py: 0.3458 seconds
ops.py: 0.2825 seconds
symbolic.py: 0.1162 seconds
helpers.py: 0.0586 seconds
```

After:

```
lazy.py: 0.5744 seconds
tensor.py: 0.4551 seconds
shapetracker.py: 0.2559 seconds
helpers.py: 0.2406 seconds
ops.py: 0.2400 seconds
symbolic.py: 0.1097 seconds
```

Deltas:

`lazy.py         -0.33s`
`shapetracker.py -0.10s`
`ops.py          -0.05s`
`helpers.py      +0.19s`

Net perf gain = `-0.29s`