@tinyb0t 2023-06-16
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/linearizer.py            17              17
tinygrad/graph.py                          7               6
tinygrad/helpers.py                       52               9
tinygrad/lazy.py                         206             153
tinygrad/mlops.py                         37              19
tinygrad/ops.py                           47              23
tinygrad/runtime/lib.py                    4               0
------------------------------------------------------------
total                                    370             227
------------------------------------------------------------
lines added in the tinygrad folder: 143
```