@rayanht 2023-06-17
Cooled down the copy paste, we lost a tiny bit of perf but the lazy ops are much more concise and readable now (apart from permute which has some optimizations that don't apply to other ops). We lost a tiny bit of perf but nothing crazy

| Filename        	| Before 	| After 	| Delta        	|
|-----------------	|--------	|-------	|--------------	|
| lazy.py         	| 0.90s  	| 0.59s 	| -0.31s       	|
| tensor.py       	| 0.45s  	| 0.44s 	| not stat-sig 	|
| ops.py          	| 0.28s  	| 0.23s 	| -0.05s       	|
| shapetracker.py 	| 0.34s  	| 0.22s 	| -0.12s       	|
| helpers.py      	| 0.05s  	| 0.22s 	| +0.17s       	|
| symbolic.py     	| 0.11s  	| 0.14s 	| +0.03s       	|
| **TOTAL**       	| 2.13s  	| 1.84s 	| -0.29s       	|