@geohot 2023-06-17
Readability of movement ops looks good.

My main concern now is the addition of `buffers` to LazyOp. This puts state in two places, both in `src` and `buffers`. Can this be a cached method instead?