@rayanht 2023-06-18
Ok about to push the changes you requested above once I make mypy happy

re: mypy, it's super upset because I tried to put actual types on the LazyOp object (it used to have `Any` for src and args) since we have no code completion otherwise. For example here the IDE has no idea that self.op.src[0] is a LazyBuffer and so it doesn't find `expand_op`. 

Now the issue with having better types is that we need to help mypy in narrowing the types but i can't think of a clean way to do that, we could:

1. Have casts and asserts everywhere, don't think this is a reasonable fix
2. Have an abstract class that defines the methods and have `LazyOp` and `LazyBuffer` both inherit it
3. Content with no code completion

<img width="561" alt="Screenshot 2023-06-19 at 00 17 12" src="https://github.com/geohot/tinygrad/assets/42040895/3cc1e222-4e94-40cf-b18e-64149884f841">