@geohot 2023-06-18
I think the abstract class approach is good, yea, I was never happy about that Any.