@rayanht 2023-06-18
Yeah it would raise a `NotImplementedError` if called on a LazyOp (which shouldn't happen by construction but that can potentially double up as an extra measure to catch breaking changes?)