@geohot 2023-06-18
Ahh, then you might not need an abstract class, perhaps add the stub methods to LazyOp and use a Union?