@rayanht 2023-06-19
Actually this only solves half of the battle, the other half is attribute accesses. We can get around that with having `@property` stubs but then it gets a bit ridiculous and we end up with something like this in the `LazyOp` class:

```
def realize(self): raise NotImplementedError
def reduce_op(self, _): raise NotImplementedError
def reshape_op(self, _): raise NotImplementedError
def pad_op(self, _): raise NotImplementedError
def expand_op(self, _): raise NotImplementedError
def permute_op(self, _): raise NotImplementedError
def shrink_op(self, _): raise NotImplementedError
def stride_op(self, _): raise NotImplementedError
@property
def children(self): raise NotImplementedError
@property
def realized(self): raise NotImplementedError
@property
def shape(self): raise NotImplementedError
@property
def st(self): raise NotImplementedError
@property
def src(self): raise NotImplementedError
[a few more of these...]
 ```

I'm not quite sure what's preferable between

1. Disabling type checking in `lazy.py` in favor of having code completion
2. Reverting to `Any` and content with no completion

My gut feeling says 1. because `lazy.py` is entirely internal APIs so we don't break anything for end-users by playing fast and loose with types and i feel like the QoL of having completion outweighs having 100% sound typing

In the long term we probably don't actually need separate classes for `LazyOp` and `LazyBuffer`, i think it can all live under `LazyBuffer` and that would solve all the typing shenanigans