@geohot 2023-06-19
I actually don't mind 25 lines of stubs. It's a good way to start thinking about merging the classes.