@rayanht 2023-06-23
Update on this, working on merging LazyOp and LazyBuffer into one, should give us a net saving in nlines and probably some additional perf as well