@geohot 2023-06-23
Cool, I'm curious to read it. My distinction was two fold.

LazyBuffer has a ShapeTracker. LazyBuffer can be realized. LazyOp doesn't and can't.