@geohot 2023-06-26
Maybe for now we should just merge this with the stubs? So it doesn't go stale.