@geohot 2023-06-26
Sweet, going to merge!

master:
```
codegen         mean runtime:  149.24ms, runs:   158.54,  146.31,  154.19,  150.24,  168.12,  143.27,  142.01,  142.79,  143.06,  143.87
methodcache     mean runtime:  141.55ms, runs:   158.67,  136.71,  136.79,  136.77,  136.66,  136.63,  165.18,  135.95,  135.75,  136.37
profile         mean runtime:  577.50ms, runs:   612.50,  595.11,  596.83,  554.52,  554.58,  562.47,  563.12,  568.43,  596.47,  571.01
```

this branch:
```
codegen         mean runtime:  103.21ms, runs:   186.60,   90.75,   90.47,   90.88,   90.44,  115.21,   91.01,   92.16,   91.96,   92.64
methodcache     mean runtime:   91.17ms, runs:    81.05,  111.85,   88.09,   88.08,   86.29,   85.49,   85.24,  113.39,   85.95,   86.26
profile         mean runtime:  362.23ms, runs:   401.66,  356.05,  346.72,  363.68,  376.81,  361.58,  352.67,  354.32,  356.23,  352.55
```