@geohot 2023-06-26
Is there more from #812?

I'm seeing 109.27 ms for the whole thing on my M1 Max