@rayanht 2023-06-26
Yeah a couple more minor things, PR coming